feat: add Drift Detection mode with payload output (#529)
Adds an output to the action that can be used to track drift detection. 

Initial intention is to write a shared workflow that can be dropped into projects to post drift to slack on a recurring basis.

Have kept it decoupled from the slack posting to make it easier to route it to different locations, for instance logging to a SIEM using the sentinel forwarder action.